refactor(classification): drop legacy AI actions, obsolete workflows, and categories/tags repos
Tasks 16+17 of the classification module migration. The new module under
modules/classification/ now owns AI classification end-to-end, so the old
implementation across @core/agents, @packages/workflows, and the categories/
tags repositories is no longer reachable.

Migrated importers before deleting the layer underneath:

- apps/web/src/integrations/orpc/router/onboarding.ts: inline
  seedEmpresarialCategories into a co-located helper at
  ./-onboarding/seed-empresarial-categories.ts (full transaction with
  WebAppError, neverthrow, no AppError).
- apps/web/src/integrations/orpc/sdk/router/categories.ts: inline list/
  create/update/remove/archive directly via context.db with WebAppError +
  fromPromise; ownership checks done inside transactions; pt-BR error
  messages.
- core/database/src/repositories/transactions-repository.ts: replace
  getCategory/getTag cross-repo calls in validateTransactionReferences
  with inline db.query.{categories,tags}.findFirst lookups.
- apps/web/src/integrations/orpc/router/transactions.ts: replace
  enqueueCategorizationWorkflow + enqueueSuggestTagWorkflow with the new
  enqueueClassifyTransactionsBatchWorkflow (collects ids per call site,
  enqueues a single batch).

Deletions:

- core/agents/src/actions/{categorize,suggest-tag,keywords,keywords-tag}.ts
  and core/agents/src/middleware/posthog.ts (439 lines). The new module's
  @core/ai/middleware + ai/classify-batch.ts replace these.
- packages/workflows/src/workflows/{categorization,suggest-tag,
  derive-keywords,derive-tag-keywords,backfill-keywords}-workflow.ts
  (924 lines).
- packages/workflows/src/workflow-factory.ts (queue keys all dead).
- core/database/src/repositories/{categories,tags}-repository.ts and
  matching tests (2751 lines).

Cleanups:

- packages/workflows/src/setup.ts no longer side-effect-imports deleted
  workflows or schedules backfill-keywords-daily (lives in
  @modules/classification/workflows/setup.ts now).
- packages/workflows/src/context.ts trimmed to bare WorkflowDeps type;
  no consumer reads getDeps()/getPublisher() anymore.
- packages/workflows/package.json drops ./workflows/* export entry.
- core/agents/package.json drops ./actions/* export entry.
- packages/notifications/src/types.ts drops AI_*/CRON_KEYWORDS_BACKFILL
  entries (only billing notification types remain).
- apps/web/src/features/notifications/use-job-notifications.ts drops
  the now-unreachable AI/cron type-specific invalidation branches.
- apps/web/README.md updated to reference the new enqueue function.

Verification:

- bun run typecheck — green workspace-wide (26 projects).
- @core/database test count drops 22→20 files (271 tests pass).
- @modules/classification still 91 tests passing.
- @modules/billing still 127 tests passing.
