+++ apps/web/src/features/notifications/use-job-notifications.ts
@@ -1,14 +1,11 @@
 import { useEffect } from "react";
-import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { toast } from "sonner";
 import { orpc } from "@/integrations/orpc/client";
-import { NOTIFICATION_TYPES } from "@packages/notifications/types";
 
 const TOAST_ID = "job-notifications";
 
 export function useJobNotifications() {
-   const queryClient = useQueryClient();
-
    const { data } = useQuery(
       orpc.notifications.subscribe.experimental_liveOptions({
          retry: true,
@@ -29,21 +26,5 @@ export function useJobNotifications() {
       }
 
       toast.info(data.message, { id: TOAST_ID });
-      if (
-         data.type === NOTIFICATION_TYPES.AI_KEYWORD_DERIVED ||
-         data.type === NOTIFICATION_TYPES.CRON_KEYWORDS_BACKFILL
-      ) {
-         queryClient.invalidateQueries({
-            queryKey: orpc.categories.getAll.queryKey(),
-         });
-      }
-      if (
-         data.type === NOTIFICATION_TYPES.AI_TRANSACTION_CATEGORIZED ||
-         data.type === NOTIFICATION_TYPES.AI_TAG_SUGGESTED
-      ) {
-         queryClient.invalidateQueries({
-            queryKey: orpc.transactions.getAll.queryKey(),
-         });
-      }
-   }, [data, queryClient]);
+   }, [data]);
 }
