+++ packages/notifications/src/types.ts
@@ -1,9 +1,4 @@
 export const NOTIFICATION_TYPES = {
-   AI_KEYWORD_DERIVED: "ai.keyword_derived",
-   AI_TRANSACTION_CATEGORIZED: "ai.transaction_categorized",
-   AI_TAG_KEYWORD_DERIVED: "ai.tag_keyword_derived",
-   AI_TAG_SUGGESTED: "ai.tag_suggested",
-   CRON_KEYWORDS_BACKFILL: "cron.keywords_backfill",
    BILLING_INVOICE_GENERATED: "billing.invoice_generated",
    BILLING_TRIAL_EXPIRING: "billing.trial_expiring",
    BILLING_BENEFIT_GRANTED: "billing.benefit_granted",
@@ -15,25 +10,6 @@ export type NotificationType =
    (typeof NOTIFICATION_TYPES)[keyof typeof NOTIFICATION_TYPES];
 
 export type NotificationPayloadMap = {
-   "ai.keyword_derived": {
-      categoryId: string;
-      categoryName: string;
-      count: number;
-   };
-   "ai.transaction_categorized": {
-      transactionId: string;
-   };
-   "ai.tag_keyword_derived": {
-      tagId: string;
-      tagName: string;
-      count: number;
-   };
-   "ai.tag_suggested": {
-      transactionId: string;
-   };
-   "cron.keywords_backfill": {
-      count: number;
-   };
    "billing.invoice_generated": {
       invoiceId: string;
       subscriptionId: string;
